+++ src/ServiceFabric/ServiceFabric/Commands/ManagedClusters/NodeTypes/RemoveAzServiceFabricManagedNodeTypeVMExtension.cs
@@ -12,20 +12,19 @@
 // ----------------------------------------------------------------------------------
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Commands.ServiceFabric.Common;
 using Microsoft.Azure.Commands.ServiceFabric.Models;
 using Microsoft.Azure.Management.Internal.Resources;
-using Microsoft.Azure.Management.ServiceFabric;
-using Microsoft.Azure.Management.ServiceFabric.Models;
+using Microsoft.Azure.Management.ServiceFabricManagedClusters;
+using Microsoft.Azure.Management.ServiceFabricManagedClusters.Models;
 
 namespace Microsoft.Azure.Commands.ServiceFabric.Commands
 {
-    [Cmdlet(VerbsCommon.Add, ResourceManager.Common.AzureRMConstants.AzurePrefix + Constants.ServiceFabricPrefix + "ManagedNodeTypeVMSecret", DefaultParameterSetName = ByObj, SupportsShouldProcess = true), OutputType(typeof(PSManagedNodeType))]
-    public class AddAzServiceFabricManagedNodeTypeVMSecret : ServiceFabricCommonCmdletBase
+    [Cmdlet(VerbsCommon.Remove, ResourceManager.Common.AzureRMConstants.AzurePrefix + Constants.ServiceFabricPrefix + "ManagedNodeTypeVMExtension", DefaultParameterSetName = ByObj, SupportsShouldProcess = true), OutputType(typeof(PSManagedNodeType))]
+    public class RemoveAzServiceFabricManagedNodeTypeVMExtension : ServiceFabricManagedCmdletBase
     {
         protected const string ByName = "ByName";
         protected const string ByObj = "ByObj";
@@ -49,24 +48,21 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Commands
         [Parameter(Mandatory = true, Position = 2, ValueFromPipelineByPropertyName = true, ParameterSetName = ByName,
             HelpMessage = "Specify the name of the node type.")]
         [ValidateNotNullOrEmpty()]
-        [Alias("NodeTypeName")]
-        public string Name { get; set; }
+        public string NodeTypeName { get; set; }
 
         [Parameter(Mandatory = true, Position = 0, ValueFromPipeline = true, ParameterSetName = ByObj,
-            HelpMessage = "Node Type resource")]
+            HelpMessage = "Node type resource")]
         [ValidateNotNull]
         public PSManagedNodeType InputObject { get; set; }
 
         #endregion
 
-        [Parameter(Mandatory = true, HelpMessage = "Key Vault resource id containing the certificates.")]
-        public string SourceVaultId { get; set; }
+        [Parameter(Mandatory = true, HelpMessage = "extension name.")]
+        [Alias("ExtensionName")]
+        public string Name { get; set; }
 
-        [Parameter(Mandatory = true, HelpMessage = "This is the URL of a certificate that has been uploaded to Key Vault as a secret. For adding a secret to the Key Vault, see [Add a key or secret to the key vault](https://docs.microsoft.com/azure/key-vault/key-vault-get-started/#add). In this case, your certificate needs to be It is the Base64 encoding of the following JSON Object which is encoded in UTF-8: <br><br> {<br>  \"data\":\"<Base64-encoded-certificate>\",<br>  \"dataType\":\"pfx\",<br>  \"password\":\"<pfx-file-password>\"<br>}/")]
-        public string CertificateUrl { get; set; }
-
-        [Parameter(Mandatory = true, HelpMessage = "Specifies the certificate store on the Virtual Machine to which the certificate should be added. The specified certificate store is implicitly in the LocalMachine account.")]
-        public string CertificateStore { get; set; }
+        [Parameter(Mandatory = false)]
+        public SwitchParameter PassThru { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background and return a Job to track progress.")]
         public SwitchParameter AsJob { get; set; }
@@ -76,17 +72,24 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Commands
         public override void ExecuteCmdlet()
         {
             this.SetParams();
-            if (ShouldProcess(target: this.Name, action: string.Format("Add Secret to node type {0}", this.Name)))
+            if (ShouldProcess(target: this.Name, action: string.Format("Remove Extenions {0} from node type {1}", this.Name, this.NodeTypeName)))
             {
                 try
                 {
-                    NodeType updatedNodeTypeParams = this.GetNodeTypeWithAddedSecret();
-                    var beginRequestResponse = this.SFRPClient.NodeTypes.BeginCreateOrUpdateWithHttpMessagesAsync(this.ResourceGroupName, this.ClusterName, this.Name, updatedNodeTypeParams)
+                    NodeType updatedNodeTypeParams = this.GetNodeTypeWithRemovedExtension();
+                    var beginRequestResponse = this.SfrpMcClient.NodeTypes.BeginCreateOrUpdateWithHttpMessagesAsync(this.ResourceGroupName, this.ClusterName, this.NodeTypeName, updatedNodeTypeParams)
                         .GetAwaiter().GetResult();
 
                     var nodeType = this.PollLongRunningOperation(beginRequestResponse);
 
-                    WriteObject(new PSManagedNodeType(nodeType), false);
+                    if (this.PassThru)
+                    {
+                        WriteObject(true);
+                    }
+                    else
+                    {
+                        WriteObject(new PSManagedNodeType(nodeType), false);
+                    }
                 }
                 catch (Exception ex)
                 {
@@ -96,41 +99,18 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Commands
             }
         }
 
-        private NodeType GetNodeTypeWithAddedSecret()
+        private NodeType GetNodeTypeWithRemovedExtension()
         {
-            NodeType currentNodeType = this.SFRPClient.NodeTypes.Get(this.ResourceGroupName, this.ClusterName, this.Name);
+            NodeType currentNodeType = this.SfrpMcClient.NodeTypes.Get(this.ResourceGroupName, this.ClusterName, this.NodeTypeName);
 
-            if (currentNodeType.VmSecrets == null)
+            if (currentNodeType.VmExtensions != null)
             {
-                currentNodeType.VmSecrets = new List<VaultSecretGroup>();
-            }
-
-            var vault = currentNodeType.VmSecrets.FirstOrDefault(v => string.Equals(v.SourceVault.Id, this.SourceVaultId, StringComparison.OrdinalIgnoreCase));
-            bool newVaultSecretGroup = false;
-            if (vault == null)
-            {
-                newVaultSecretGroup = true;
-                vault = new VaultSecretGroup()
+                var originalLength = currentNodeType.VmExtensions.Count();
+                currentNodeType.VmExtensions = currentNodeType.VmExtensions.Where(ext => !string.Equals(ext.Name, this.Name, StringComparison.OrdinalIgnoreCase)).ToList();
+                if (originalLength == currentNodeType.VmExtensions.Count())
                 {
-                    SourceVault = new SubResource(this.SourceVaultId)
-                };
-            }
-            
-            if (vault.VaultCertificates == null)
-            {
-                vault.VaultCertificates = new List<VaultCertificate>()
-                {
-                    new VaultCertificate()
-                    {
-                        CertificateStore = this.CertificateStore,
-                        CertificateUrl = this.CertificateUrl
-                    }
-                };
-            }
-
-            if (newVaultSecretGroup)
-            {
-                currentNodeType.VmSecrets.Add(vault);
+                    throw new ArgumentException(string.Format("extension with name {0} not found", this.Name));
+                }
             }
 
             return currentNodeType;
@@ -148,7 +128,6 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Commands
 
                     SetParametersByResourceId(this.InputObject.Id);
                     break;
-
             }
         }
 
@@ -156,7 +135,7 @@ namespace Microsoft.Azure.Commands.ServiceFabric.Commands
         {
             this.GetParametersByResourceId(resourceId, Constants.ManagedNodeTypeProvider, out string resourceGroup, out string resourceName, out string parentResourceName);
             this.ResourceGroupName = resourceGroup;
-            this.Name = resourceName;
+            this.NodeTypeName = resourceName;
             this.ClusterName = parentResourceName;
         }
     }